move function to common db utils
